requirements.txt
# 注意：torch-geometric 必须使用 == 锁定版本，否则会自动升级 PyTorch！
# 必须先安装torch==2.1.0，然后安装torch-geometric，避免torch被升级
torch==2.1.0
torch-geometric==2.4.0

ninja
easydict
pyyaml
huggingface_hub==0.19.4
transformers==4.39.3
sentence-transformers==2.2.2
numpy==1.24.1
einops
# flash-attn # Note: flash-attn installation can be complex and may require specific CUDA toolkit versions.
python-dotenv
httpx
aiohttp
nltk